Cleanup: separate WAPI and Drive API code in gdata_documents_service.
BUG=142809
TEST=pass all existing tests, manually check if file browser doesn't break.
TBR=ben@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10829375

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@152476 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 02782e2e9fa76b67ad971fd19c3eb1bc0d17ca57

+++ chrome/browser/chromeos/gdata/mock_documents_service.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "chrome/browser/chromeos/gdata/mock_gdata_documents_service.h"
+#include "chrome/browser/chromeos/gdata/mock_documents_service.h"
 
 #include "base/bind.h"
 #include "base/file_path.h"
@@ -90,7 +90,7 @@ void MockDocumentsService::AuthenticateStub(
 
 void MockDocumentsService::GetDocumentsStub(
     const GURL& feed_url,
-    int start_changestamp,
+    int64 start_changestamp,
     const std::string& search_string,
     const std::string& directory_resource_id,
     const GetDataCallback& callback) {
